+++ src/Components/Home/Home.js
@@ -1,6 +1,6 @@
 import "../../Css/home.css";
 
-// Returns the main page of the website
+/* This component returns the main page of the website. */
 
 const Main = () => {
   return (