+++ src/Components/ProtectedRoute/ProtectedRoute.js
@@ -3,6 +3,8 @@ import { useNavigate } from "react-router-dom";
 import { checkUser } from "../../Common/Services/AuthService";
 import "../../Css/protectedRoute.css";
 
+/* This component makes sure unauthenticated users cannot visit protected routes. */
+
 const ProtectedRoute = ({ element: Component, ...rest }) => {
   console.log("element: ", Component);
   const navigate = useNavigate();